Chunk commits by date in lists
It's possible to construct a commit graph where the output of `git log`
isn't in timestamp order. Grouping the commits in the list by date then
gives dramatically wrong results. Instead, go for the more pragmatic
approach: use the commits in the order they're given, and just show the
date line each time the date changes. This means that the same date
header can show up multiple times, but at least the ordering is
preserved.