Replace `link_to_gfm` with `link_to` in merge request and issue titles.
Previously if an issue had an emoji for its name it would be unclickable. Similarly, links were rendered inline if the title was something like “Fixes a bug with https://google.com”. This confused some users, so it’s removed.

On the issue/MR page itself, the formatting is preserved. This only effects index pages.

This is also ever so slightly more performant :)

This fixes #17614 and #17230.