Extra tests inside shared_examples:
Feedback from:
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/4093#note_12321421

The advantage of this is that tests are less about logic and more
straightforward, thus could be easier to reason about each individual
tests.

The disadvantage of this is that we write more duplicated codes and
once something changed we might need to change all places and it's
harder to reason all tests as a whole. Because now we need to look at
more places to figure out how it should work under another option!