Add a small locked icon if it's locked:
This is probably not the way we add icons, but well. Not sure if this
should be put next to the status icon or edit icon, my first thought
was put this next to status, but it looks a bit better next to edit
button. Please tweak this accordingly. I don't have strong opinion
regarding views.

Feedback:
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/4093#note_12345567