Merge branch 'diff-commit-id-weight' into 'master'
Fixes font weight of commit id on view file button

## What does this MR do?

Changes the weight to normal. With the new commit UI - this had turned bold and didn't match the designs.

Should it also be our normal font family? According to designs it is - but everywhere else there is a commit ID it has the monospace font.

## Screenshots (if relevant)

![Screen_Shot_2016-06-16_at_17.25.33](/uploads/bb76d576c10be65172d5f4e9790ca766/Screen_Shot_2016-06-16_at_17.25.33.png)

See merge request !4713