Update wordings accordingly because:
It's not that it must be rebased, it could also be merged.
This also raises a question: Why we could only rebase on
GitLab, but not merge whenever merge should also be an option?

After this change, we could also not enforce a linear history
because we also allow merging.