Add some clarification for some files under config/*
I didn't add a lot of them because I am not very familiar about how they
work and I am not sure if we really need to explain files like
config/database.yml because if you know Rails you must know how
that works.

At any rate, I think this is a beginning and we could keep documenting
more files in the future.

Closes #18373