Have CI test migration paths
This commit adds a job to the pipeline to test if migration on seed
data works. Even though the seed data is not perfect, it does give more
guarentees that this will succeed in the real world too.

seed_fu is used instead of dev:setup to save some time, as they both
generate the data in the same way, this should be fine.