Disable the email checking part of the standard Health Check
The email check used in the Heath Check doesn't properly make use of enough of the SMTP config options
to be able to properly test the STMP connection, and as a result could cause a failure.

In order to fix it we have overwritten the email_configured? method in the health check
so that it does not check email status during the standard health check.