Remove the group avatar link.
For group pages, e.g. https://gitlab.com/groups/gitlab-org, I see no reason to link to the avatar image itself when clicking on the avatar.

If the user really wants to open the image, they can do so by right clicking and opening the image in a new tab.