Merge branch 'api-dev-can-push' into 'master'
API: Expose 'developers_can_push' for branches

## What does this MR do?

Adds support for the `developers_can_push` flag for the branches API. It also supports creating protected branches with that flag.

## Are there points in the code the reviewer needs to double check?

The API call requires an optional boolean parameter `developers_can_push`. If it is not either `true` or `false`, it should not update the value. I'm not sure if this is the right way to do it.

Right now it always returns `200`. Maybe the `PUT` method should return different status codes depending if it really created or updated a branch.

## What are the relevant issue numbers?

Closes #12735 

## Does this MR meet the acceptance criteria?

- [x] [CHANGELOG](https://gitlab.com/gitlab-org/gitlab-ce/blob/master/CHANGELOG) entry added
- [x] [Documentation created/updated](https://gitlab.com/gitlab-org/gitlab-ce/blob/master/doc/development/doc_styleguide.md)
- [x] API support added
- Tests
  - [x] Added for this feature/bug
  - [x] All builds are passing
- [x] Conform by the [style guides](https://gitlab.com/gitlab-org/gitlab-ce/blob/master/CONTRIBUTING.md#style-guides)
- [x] Branch has no merge conflicts with `master` (if you do - rebase it please)
- [x] [Squashed related commits together](https://git-scm.com/book/en/Git-Tools-Rewriting-History#Squashing-Commits)

See merge request !5208