Merge branch 'registry-manifest' into 'master'
Document when Docker Registry manifest v1 support was added

## What does this MR do?

Document when Docker Registery manifest v1 support was added

## Are there points in the code the reviewer needs to double check?

Is this the best way to document this? We never documented support for v2 or which docker versions we support overall.

## Why was this MR needed?

We made a change and it wasn't documented. Also, it's good to document which docker versions we support.

## What are the relevant issue numbers?

!4669

## Screenshots (if relevant)


See merge request !5378