Fix expand all diffs button in compare view
We can't reuse the existing value of the format parameter, because on
the merge request page that's 'json', so the link would go to a JSON
file. We can't set it to HTML, because that adds '.html' the URL, which
breaks the compare view (and looks bad). Setting it to nil uses the
default format, which in all of these cases is HTML anyway.