Merge branch '20241-expand-all-button-breaks-diff-comparison-view' into 'master'
Fix expand all diffs button in compare view

## What does this MR do?

Fix the expand all diffs button in compare view.

## Are there points in the code the reviewer needs to double check?

Don't think so.

## Why was this MR needed?

We can't reuse the existing value of the format parameter, because on
the merge request page that's 'json', so the link would go to a JSON
file. We can't set it to HTML, because that adds '.html' to the URL, which
breaks the compare view (and looks bad). Setting it to `nil` uses the
default format, which in all of these cases is HTML anyway.

## What are the relevant issue numbers?

Closes #20241.

See merge request !5500

Signed-off-by: Rémy Coutable <remy@rymai.me>